refactor import view further
