+++ app/assets/javascripts/dashboard/dataset/dataset_import_view.js
@@ -19,7 +19,6 @@ import {
   Switch,
   Tooltip,
 } from "antd";
-import Clipboard from "clipboard-js";
 import update from "immutability-helper";
 import Toast from "libs/toast";
 import {
@@ -33,7 +32,6 @@ import {
   updateDatasetDatasource,
   updateDatasetTeams,
 } from "admin/admin_rest_api";
-import { Vector3Input, BoundingBoxInput } from "libs/vector_input";
 import type { DatasetConfigurationType } from "oxalis/store";
 import messages from "messages";
 import type {
@@ -41,9 +39,7 @@ import type {
   APIMessageType,
   APIDataSourceWithMessagesType,
 } from "admin/api_flow_types";
-import TeamSelectionComponent from "dashboard/dataset/team_selection_component";
 import { handleGenericError } from "libs/error_handling";
-import { validateLayerConfigurationJSON } from "./validation";
 import {
   Hideable,
   FormItemWithInfo,
@@ -52,6 +48,8 @@ import {
   jsonEditStyle,
 } from "./helper_components";
 import SimpleAdvancedDataForm from "./simple_advanced_data_form";
+import DefaultConfigComponent from "./default_config_component";
+import ImportGeneralComponent from "./import_general_component";
 
 const FormItem = Form.Item;
 const TabPane = Tabs.TabPane;
@@ -77,7 +75,7 @@ type State = {
   activeTabKey: TabKeyType,
 };
 
-type FormData = {
+export type FormData = {
   dataSource: APIDataSourceWithMessagesType,
   dataSourceJson: string,
   dataset: APIDatasetType,
@@ -270,11 +268,6 @@ class DatasetImportView extends React.PureComponent<Props, State> {
     });
   };
 
-  handleCopySharingLink = async () => {
-    await Clipboard.copy(this.getSharingLink());
-    Toast.success("Sharing Link copied to clipboard");
-  };
-
   handleRevokeSharingLink = async () => {
     this.setState({ isLoading: true });
     try {
@@ -286,21 +279,11 @@ class DatasetImportView extends React.PureComponent<Props, State> {
     }
   };
 
-  handleSelectText = (event: SyntheticInputEvent<>) => {
-    event.target.select();
-  };
-
   updateDataset(propertyName: string, value: string | boolean) {
     const newState = update(this.state, {
       dataset: { [propertyName]: { $set: value } },
     });
     this.setState(newState);
-  }
-
-  getSharingLink() {
-    return `${window.location.origin}/datasets/${this.props.datasetName}/view?token=${
-      this.state.sharingToken
-    }`;
   }
 
   getMessageComponents() {
@@ -329,152 +312,6 @@ class DatasetImportView extends React.PureComponent<Props, State> {
     return (
       this.props.form.getFieldValue("dataset.allowedTeams") == null ||
       this.props.form.getFieldValue("dataset.allowedTeams").length === 0
-    );
-  }
-
-  getGeneralComponents() {
-    const { getFieldDecorator } = this.props.form;
-    const _hasNoAllowedTeams = this.hasNoAllowedTeams();
-
-    const allowedTeamsComponent = (
-      <FormItemWithInfo
-        label="Allowed Teams"
-        info="Except for administrators and team managers, only members of the teams defined here will be able to view this dataset."
-        validateStatus={_hasNoAllowedTeams ? "warning" : "success"}
-        help={
-          _hasNoAllowedTeams
-            ? "If this field is empty, only administrators and team managers will be able to view this dataset."
-            : null
-        }
-      >
-        {getFieldDecorator("dataset.allowedTeams", {})(<TeamSelectionComponent mode="multiple" />)}
-      </FormItemWithInfo>
-    );
-    const content = !this.props.isEditingMode ? (
-      allowedTeamsComponent
-    ) : (
-      <div>
-        <Row gutter={48}>
-          <Col span={12}>
-            <FormItemWithInfo
-              label="Display Name"
-              info="The display name will be used by webKnossos to name this dataset. If the display name is not provided, the original name will be used."
-            >
-              {getFieldDecorator("dataset.displayName")(<Input placeholder="Display Name" />)}
-            </FormItemWithInfo>
-          </Col>
-          <Col span={12}>
-            <FormItemWithInfo
-              label="Description"
-              info="The description may contain additional information about your dataset."
-            >
-              {getFieldDecorator("dataset.description")(
-                <Input.TextArea rows="3" placeholder="Description" />,
-              )}
-            </FormItemWithInfo>
-          </Col>
-        </Row>
-        {allowedTeamsComponent}
-        <FormItemWithInfo
-          label="Sharing Link"
-          info="The sharing link can be used to allow unregistered users viewing this dataset."
-        >
-          <Input.Group compact>
-            <Input
-              value={this.getSharingLink()}
-              onClick={this.handleSelectText}
-              style={{ width: "80%" }}
-              readOnly
-            />
-            <Button onClick={this.handleCopySharingLink} style={{ width: "10%" }} icon="copy" />
-            <Button onClick={this.handleRevokeSharingLink} style={{ width: "10%" }}>
-              Revoke
-            </Button>
-          </Input.Group>
-        </FormItemWithInfo>
-        <FormItem>
-          {getFieldDecorator("dataset.isPublic", { valuePropName: "checked" })(
-            <Checkbox>
-              Make dataset publicly accessible{" "}
-              <Tooltip title="If checked, the dataset will be listed when unregistered users visit webKnossos.">
-                <Icon type="info-circle-o" style={{ color: "gray" }} />
-              </Tooltip>
-            </Checkbox>,
-          )}
-        </FormItem>
-      </div>
-    );
-
-    return content;
-  }
-
-  getDefaultConfigComponents() {
-    const { getFieldDecorator } = this.props.form;
-
-    return (
-      <div>
-        <Alert
-          message="The following settings define the default configuration when viewing or creating an explorational tracing for this dataset. Use them to optimize the initial appearance of your dataset."
-          type="info"
-          showIcon
-        />
-        <Row gutter={24}>
-          <Col span={6}>
-            <FormItemWithInfo
-              label="Position"
-              info="The default position defined is defined in voxel-coordinates (x, y, z)."
-            >
-              {getFieldDecorator("defaultConfiguration.position")(<Vector3Input />)}
-            </FormItemWithInfo>
-          </Col>
-          <Col span={6}>
-            <FormItemWithInfo
-              label="Zoom"
-              info="A zoom of &ldquo;1&rdquo; will display the data in its original resolution."
-            >
-              {getFieldDecorator("defaultConfiguration.zoom")(
-                <InputNumber style={{ width: "100%" }} />,
-              )}
-            </FormItemWithInfo>
-          </Col>
-          <Col span={6}>
-            <FormItemWithInfo
-              label="Segmentation Opacity"
-              info="The segmentation layer will be overlayed using the specified percentage value (&ldquo;20&rdquo; means &ldquo;20%&rdquo; opacity)."
-            >
-              {getFieldDecorator("defaultConfiguration.segmentationOpacity")(
-                <InputNumber style={{ width: "100%" }} />,
-              )}
-            </FormItemWithInfo>
-          </Col>
-          <Col span={6}>
-            <FormItem label=" " colon={false}>
-              {getFieldDecorator("defaultConfiguration.interpolation", {
-                valuePropName: "checked",
-              })(
-                <Checkbox>
-                  Interpolation{" "}
-                  <Tooltip
-                    title={
-                      "If checked, bilinear interpolation will be used the the data is rendered."
-                    }
-                  >
-                    <Icon type="info-circle-o" style={{ color: "gray" }} />
-                  </Tooltip>
-                </Checkbox>,
-              )}
-            </FormItem>
-          </Col>
-        </Row>
-        <FormItemWithInfo
-          label="Layer Configuration"
-          info="Use the following JSON to define layer-specific properties, such as color, contrast and brightness."
-        >
-          {getFieldDecorator("defaultConfigurationLayersJson", {
-            rules: [{ validator: validateLayerConfigurationJSON }],
-          })(<Input.TextArea rows="10" style={jsonEditStyle} />)}
-        </FormItemWithInfo>
-      </div>
     );
   }
 
@@ -511,7 +348,8 @@ class DatasetImportView extends React.PureComponent<Props, State> {
         <Icon type="exclamation-circle" style={{ color: "#f5222d", marginLeft: 4 }} />
       </Tooltip>
     );
-    const hasNoAllowedTeamsWarning = this.hasNoAllowedTeams() ? (
+    const _hasNoAllowedTeams = this.hasNoAllowedTeams();
+    const hasNoAllowedTeamsWarning = _hasNoAllowedTeams ? (
       <Tooltip title="Please double-check some fields here.">
         <Icon type="exclamation-circle" style={{ color: "#faad14", marginLeft: 4 }} />
       </Tooltip>
@@ -562,7 +400,15 @@ class DatasetImportView extends React.PureComponent<Props, State> {
                   key="general"
                   forceRender
                 >
-                  {this.getGeneralComponents()}
+                  <ImportGeneralComponent
+                    form={form}
+                    hasNoAllowedTeams={_hasNoAllowedTeams}
+                    sharingLink={`${window.location.origin}/datasets/${
+                      this.props.datasetName
+                    }/view?token=${this.state.sharingToken}`}
+                    handleRevokeSharingLink={this.handleRevokeSharingLink}
+                    isEditingMode={this.props.isEditingMode}
+                  />
                 </TabPane>
                 {this.props.isEditingMode ? (
                   <TabPane
@@ -572,7 +418,7 @@ class DatasetImportView extends React.PureComponent<Props, State> {
                     key="defaultConfig"
                     forceRender
                   >
-                    {this.getDefaultConfigComponents()}
+                    <DefaultConfigComponent form={form} />
                   </TabPane>
                 ) : null}
               </Tabs>
